feat: Implemented a base of IGrabber.
