+++ Assets/Conekton/ARUtility/Scripts/UseCase/HandGrabSystem/Infrastructure/Grabber.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8a91e790681bfb048beff5ca6215e2bb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
